requirements.txt
asgiref==3.7.2
Django==4.2.2
flake8==7.0.0
mccabe==0.7.0
pycodestyle==2.11.1
pyflakes==3.2.0
sqlparse==0.4.4
typing_extensions==4.10.0
djangorestframework==3.14.0
psycopg2==2.9.9
drf-spectacular==0.27.1
